# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AbsenKu API
  version: 1.0.0
  description: |
    Spesifikasi kontrak API untuk frontend dan backend AbsenKu.

    Peran pengguna: SUPER_ADMIN, TEACHER, dan HOMEROOM_TEACHER.
    Semua waktu memakai ISO 8601; zona sekolah Asia/Jakarta.
servers:
  - url: https://attendly-api-three.vercel.app/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/google:
    post:
      tags: [Auth]
      summary: Masuk menggunakan Google
      operationId: loginWithGoogle
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenPairResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Perbarui access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenPairResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Token tidak valid atau kedaluwarsa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    BadRequest:
      description: Payload tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, message, error]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
    AuthTokenPairResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          $ref: '#/components/schemas/AuthTokenPair'
...
    Uuid:
      type: string
      format: uuid
...
    Role:
      type: string
      enum: [SUPER_ADMIN, TEACHER, HOMEROOM_TEACHER]
...
    GoogleLoginRequest:
      type: object
      additionalProperties: false
      required: [id_token]
      properties:
        id_token:
          type: string
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    CurrentUser:
      type: object
      required: [id, email, name, roles, permissions]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        permissions:
          type: array
          items:
            type: string
    AuthTokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/CurrentUser'
...